pyproject.toml
[project]
name = "ntd-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "faker>=38.2.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "I"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ntd_challenge.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
filterwarnings = [
    "ignore::django.utils.deprecation.RemovedInDjango60Warning"
]
